+++ crates/adapters/blockchain/src/rpc/core.rs
@@ -206,6 +206,8 @@ impl CoreBlockchainRpcClient {
                 .proxy_url
                 .as_ref()
                 .map(|value| value.expose_secret().to_owned()),
+            max_message_size_bytes: None,
+            max_frame_size_bytes: None,
         };
 
         let client = WebSocketClient::builder()